backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.9
python-jose[cryptography]==3.5.0
cryptography==44.0.2
httpx==0.28.1
python-multipart==0.0.20
